==> analysis/01-LoadPancreasData.Rmd <==
---
title: "Load single-cell data"
author: "Nicolas Damond and Nils Eling"
date: "`r Sys.Date()`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script downloads the single-cell data from the pancreas IMC dataset available [here](http://dx.doi.org/10.17632/cydmwsfztj.2).
The dataset is associated to the following publication:

[Damond et al. A Map of Human Type 1 Diabetes Progression by Imaging Mass Cytometry. Cell Metabolism. 2019 Mar 5;29(3):755-768](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6821395)

After obtaining the raw data as produced by the [IMC segmentation pipeline](https://github.com/BodenmillerGroup/ImcSegmentationPipeline), we will further process them to create a [SingleCellExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html) object.

# Obtain and read-in data

Here, a subset of single-cell data, corresponding to 100 images from the full dataset is downloaded.

## Read in single-cell data

We read in the single-cell meta- and expression data and order them based on the image and cell number.

```{r download-cell-data, message=FALSE}
library(S4Vectors)
library(SingleCellExperiment)
library(cytomapper)

# Download the zipped folder image and unzip it
url.cells <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/f1e3b8dc-56be-4172-bbc4-3a6f9de97563/file_downloaded")
download.file(url.cells, destfile = "data/PancreasData/CellSubset.zip")
unzip("data/PancreasData/CellSubset.zip", exdir = "data/PancreasData")
file.remove("data/PancreasData/CellSubset.zip")

# Read-in the data
cells <- read.csv("data/PancreasData/CellSubset.csv", stringsAsFactors = FALSE)

# Order the dataset by ImageNumber and ObjectNumber
cells <- cells[order(cells$ImageNumber, cells$ObjectNumber), ]
```

## Read in image metadata

Next, we will read in the image-specific metadata.

```{r download-image-metadata}
# Download the zipped folder image and unzip it
url.image <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/0b236273-d21b-4566-84a2-f1c56324a900/file_downloaded")
download.file(url.image, destfile = "data/PancreasData/Image.zip")
unzip("data/PancreasData/Image.zip", exdir = "data/PancreasData")
file.remove("data/PancreasData/Image.zip")

# Read-in the data
image <- read.csv("data/PancreasData/All_Image.csv", stringsAsFactors = FALSE)
```

## Read in cell-type information

In the original publication, cells were phenotyped based on informative marker expression.
These phenotype labels are supplied in the online repository.

```{r download-celltypes}
# Download the zipped folder image and unzip it
url.celltypes <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/59e8da72-5bfe-4289-b95b-28348a6e1222/file_downloaded")
download.file(url.celltypes, destfile = "data/PancreasData/CellTypes.zip")
unzip("data/PancreasData/CellTypes.zip", exdir = "data/PancreasData")
file.remove("data/PancreasData/CellTypes.zip")

# Read-in the data
celltypes <- read.csv("data/PancreasData/CellTypes.csv", stringsAsFactors = FALSE)
```

## Read-in donor information

We will furthermore read in the metadata per donor.

```{r download-donor-info}
# Download the zipped folder image and unzip it
url.donors <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/9074990e-1b93-4c79-8c49-1db01a66398b/file_downloaded")
download.file(url.donors, destfile = "data/PancreasData/Donors.zip")
unzip("data/PancreasData/Donors.zip", exdir = "data/PancreasData")
file.remove("data/PancreasData/Donors.zip")

# Read-in the data
donors <- read.csv("data/PancreasData/Donors.csv", stringsAsFactors = FALSE)
```

# Prepare metadata

In this part of the workflow, we will select cell-specific metadata and collect them in a single `DataFrame`.
This will be the `colData` entry of the final `SingleCellExperiment` object.

## Generate cell-specific metadata

The cell-specific metadata can be obtained from the `cells` object.

```{r cell-metadata}
cell.metadata <- DataFrame(ImageNumber = cells$ImageNumber,
                           CellNumber = cells$ObjectNumber,
                           Pos_X = cells$Location_Center_X,
                           Pos_Y = cells$Location_Center_Y,
                           ParentIslet = cells$Parent_Islets,
                           ClosestIslet = cells$Parent_ExpandedIslets,
                           Area = cells$AreaShape_Area,
                           NbNeighbours = cells$Neighbors_NumberOfNeighbors_3)
```

Specific image metadata can be obtained from column-entries of the `image` object.

```{r image-metadata}
image.metadata <- DataFrame(ImageNumber = image$ImageNumber,
                            ImageFullName = image$FileName_CleanStack,
                            slide = image$Metadata_Slide,
                            width = image$Width_CleanStack,
                            height = image$Height_CleanStack)
```

We will next merge the cell- and image-specific metadata.

```{r merge-metadata}
cell.metadata <- merge(cell.metadata, image.metadata, by="ImageNumber")
```

We will also add the `ImageName` entry to the metadata object.
This information is used by `cytomapper` to match single-cell data with images and masks.

```{r add-image-names}
cell.metadata$ImageName <- sub("_a0_full_clean.tiff", "", cell.metadata$ImageFullName)
```

We will now add the cell-type information to the metadata object.

```{r import-cell-types}
# Add cell ids to cell metadata (format: "ImageName_CellNumber")
cell.metadata$id <- paste(cell.metadata$ImageName, cell.metadata$CellNumber, sep="_")

# Merge cell metadata and cell type information
cell.metadata <- merge(cell.metadata,
                       celltypes[, c("id", "CellCat", "CellType")],
                       by="id")
```

We will add the donor information to the metadata object.

```{r import-donors-info}
cell.metadata <- merge(cell.metadata, donors, by="slide")
```

Finally, we order the cell-metadata object based on `ImageNumber` and `CellNumber` and add rownames.

```{r cell-rownames}
# Rows are ordered by ImageNumber and CellNumber
cell.metadata <- cell.metadata[order(cell.metadata$ImageNumber, cell.metadata$CellNumber), ]

# Cell ids are used as row names
rownames(cell.metadata) <- cell.metadata$id
```

## Generate marker-specific metadata

Here, we will download the panel information, which contains antibody-related metadata.
However, for some datasets, the channel-order and the panel order do not match.
For this, the channel-mass file is used to match panel information and image stack slices.

```{r load-panel}
# Import panel
url.panel <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/2f9fecfc-b98f-4937-bc38-ae1b959bd74d/file_downloaded")
download.file(url.panel, destfile = "data/PancreasData/panel.csv")
panel <- read.csv("data/PancreasData/panel.csv")

# Import channel-mass file
url.channelmass <- ("https://data.mendeley.com/public-files/datasets/cydmwsfztj/files/704312eb-377c-42e2-8227-44bb9aca0fb3/file_downloaded")
download.file(url.channelmass, destfile = "data/PancreasData/ChannelMass.csv")
channel.mass <- read.csv("data/PancreasData/ChannelMass.csv", header = FALSE)
```

First, we will select the relevant channels used for analysis and order the pannel based on channel mass.

```{r arrange-panel}
# Match panel and stack slice information
panel <- panel[panel$full == 1,]
panel <- panel[match(channel.mass[,1], panel$MetalTag),]

# Add short protein names as panel rownames
rownames(panel) <- panel$shortname
```

# Select relevant expression features

CellProfiler measures a number of different statistics per marker and cell.
We will select the mean intensity per channel and per cell to obtain single-cell expression counts.

```{r select-counts}
cur_counts <- cells[, grepl("Intensity_MeanIntensity_CleanStack", colnames(cells))]
```

Next, we will reorder the channels based on channel number.

```{r reorder-counts}
channelNumber <- as.numeric(sub("^.*_c", "", colnames(cur_counts)))
cur_counts <- cur_counts[, order(channelNumber, decreasing = FALSE)]
```

# Create the SingleCellExperiment object

We have now obtained all metadata and feature data to create the `SingleCellExperiment` object.
We will first create it is based on the raw expression counts.

```{r create-SCE}
sce <- SingleCellExperiment(assays = list(counts = t(as.matrix(cur_counts))))
```

Furthermore, we will store the arcsinh-transformed (using a co-factor of 1) counts in the `exprs` assay slot.

```{r transform-counts}
assay(sce, "exprs") <- asinh(counts(sce)/1)
```

Now, we will set the dimnames of the object.

```{r dimnames-SCE}
rownames(sce) <- rownames(panel)
colnames(sce) <- rownames(cell.metadata)
```

Finally, we will store the marker- and cell-specific metadata in the `SingleCellExperiment` object.
Here, columns are cells and rows are markers.

```{r store-metadata}
colData(sce) <- cell.metadata
rowData(sce) <- panel
sce
```

# Save SCE

For further analysis, we will save the `SingleCellExperiment` object.

```{r save-SCE}
saveRDS(sce, "data/PancreasData/pancreas_sce.rds")
```

# Clean up

Finally, we remove the downloaded objects to save storage space.

```{r clean-up, message = FALSE}
file.remove("data/PancreasData/All_Image.csv", 
            "data/PancreasData/CellSubset.csv", 
            "data/PancreasData/CellTypes.csv", 
            "data/PancreasData/Donors.csv", 
            "data/PancreasData/panel.csv", 
            "data/PancreasData/ChannelMass.csv")
```
