==> analysis/03-Figures.Rmd <==
---
title: "Figures"
author: "Nils Eling and Nicolas Damond"
date: "2020-06-02"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script reproduces the analysis performed in the `cytomapper` publication and generates all displayed figures.
The analysis was performed using cytomapper version 1.0.0, Bioconductor version 3.11 and R version 4.0.

Here, we will load the libraries required for data analysis:

```{r load-libraries, message=FALSE}
library(dplyr)
library(SingleCellExperiment)
library(cytomapper)
```

## Figure 1

In the first figure, we will solely display the example data shipped with `cytomapper`.
They can be read in using the `data` function.

```{r load-example-data}
data("pancreasSCE")
data("pancreasImages")
data("pancreasMasks")
```

We will save the segmentation masks and single-channel images for image 1 and 3 doe display purposes.

```{r save-example-data}
plotCells(pancreasMasks[1], legend = NULL, image_title = NULL, 
          scale_bar = NULL, interpolate = FALSE, 
          save_plot = list(filename = "docs/final_figures/main/Fig_1A_mask1.png"))
plotCells(pancreasMasks[3], legend = NULL, image_title = NULL, 
          scale_bar = NULL, interpolate = FALSE, 
          save_plot = list(filename = "docs/final_figures/main/Fig_1A_mask2.png"))

for(i in 1:5){
  cur_col <- list(c("black", "white"))
  names(cur_col) <- channelNames(pancreasImages)[i]
  plotPixels(pancreasImages[1], legend = NULL, image_title = NULL, 
          scale_bar = NULL, interpolate = FALSE, colour_by = channelNames(pancreasImages)[i],
          colour = cur_col,
          save_plot = list(filename = paste0("docs/final_figures/main/Fig_1B_img1_ch", i, ".png")))
  
    plotPixels(pancreasImages[3], legend = NULL, image_title = NULL, 
          scale_bar = NULL, interpolate = FALSE, colour_by = channelNames(pancreasImages)[i],
          colour = cur_col,
          save_plot = list(filename = paste0("docs/final_figures/main/Fig_1B_img2_ch", i, ".png")))
}
```

We will next visualize the example cell-types on the segmentation masks and the composite images of 5 colours.

```{r viz-example-images}
plotCells(mask = pancreasMasks[1], object = pancreasSCE, img_id = "ImageNb", cell_id = "CellNb",
          colour_by = "CellType", 
          save_plot = list(filename = "docs/final_figures/main/Fig_1A_celltype1.png",
                           scale = 10))
plotCells(mask = pancreasMasks[3], object = pancreasSCE, img_id = "ImageNb", cell_id = "CellNb",
          colour_by = "CellType", 
          save_plot = list(filename = "docs/final_figures/main/Fig_1A_celltype2.png",
                           scale = 10))

plotPixels(image = pancreasImages[1], object = pancreasSCE,
           img_id = "ImageNb", cell_id = "CellNb", colour_by = channelNames(pancreasImages),
           bcg = list(H3 = c(0, 4, 1),
                      CD99 = c(0, 3, 1),
                      PIN = c(0, 5, 1),
                      CD8a = c(0, 5, 1),
                      CDH = c(0, 3, 1)),
           save_plot = list(filename = "docs/final_figures/main/Fig_1B_composite1.png",
                           scale = 10))
plotPixels(image = pancreasImages[3], object = pancreasSCE,
           img_id = "ImageNb", cell_id = "CellNb", colour_by = channelNames(pancreasImages),
           bcg = list(H3 = c(0, 2, 1),
                      CD99 = c(0, 3, 1),
                      PIN = c(0, 1, 1),
                      CD8a = c(0, 4, 1),
                      CDH = c(0, 3, 1)),
           save_plot = list(filename = "docs/final_figures/main/Fig_1B_composite2.png",
                           scale = 10))
```

## Figure 2

The previous scripts have generated the following files in the `data/PancreasData` folder:

1. A `SingleCellExperiment` object containing the cell- and marker-specific expression and metadata.
2. A `CytoImageList` object containing 100 example images of 3 donors, where each image contains 38 channels
3. A `CytoImageList` object containing 100 example masks, which are associated to and a segmentation result of the images in 2.

Here, we will load all relevant libraries and read in the data.

```{r read-in-data, message=FALSE}
sce <- readRDS("data/PancreasData/pancreas_sce.rds")
images <- readRDS("data/PancreasData/pancreas_images.rds")
masks <- readRDS("data/PancreasData/pancreas_masks.rds")
```

In type 1 diabetes (T1D), pancreatic insulin-producing beta cells are killed by self-reacting immune cells. 
Here, we will visualize association between T cells and beta cells in three pancreas donors at different stage of the disease (Non-diabetic, recent onset T1D and long duration T1D).  

In our dataset, cell types can be visualized in the `CellType` column of `colData(sce)` (`unique(colData(sce)$CellType)`). 
T cells have the cell type `Tc` (CD8+) or `Th` (CD4+) and beta cells are labelled as `beta`.  

First, we will select for each donor the image with the highest density of T cells (number of T cells per mm^2). 
Then, we will subset the `SingleCellExperiment` object (`sce`) and the `CytoImageList` objects (`images` and `masks`).

```{r select-images}
# Select the three images with the higest T cell density
selected.images <- as_tibble(colData(sce)) %>% 
  # Calculate for each image the area, number of T cells and T cell density
  group_by(ImageNumber) %>%
  mutate(ImageArea = (width * height) / 10^6,
         TcellCount = sum(CellType == "Tc" | CellType == "Th"),
         TcellDensity = TcellCount / ImageArea) %>%
  ungroup() %>%
  # Select for each disease stage the image with the highest T cell density
  group_by(stage) %>%
  dplyr::slice(which.max(TcellDensity)) %>%
  pull(ImageName)

print(selected.images)

# Subset the objects
cur_sce <- sce[, sce$ImageName %in% selected.images]
cur_images <- images[mcols(images)$ImageName %in% selected.images]
cur_masks <- masks[mcols(masks)$ImageName %in% selected.images]
```

## Plot cells

To visualize association between islet and T cells, we will use the `plotCells` function and display beta and T cells on the three images selected above.  

The different cell types will be coloured as following: Tc cells in red, Th cells in blue, beta cells in yellow, the other islet cells (alpha, delta) in cyan and the rest of the cells in black.

To display only a selected number of cells, the `SingleCellExperiment` can be subsetted.

```{r plotCells}
cur_sce <- cur_sce[,cur_sce$CellType %in% c("beta", "alpha", "delta", "Tc", "Th")]

# Define the colors for the different cell types
ct_colours <- vector(mode = "character", length = length(unique(cur_sce$CellType)))
names(ct_colours) <- unique(cur_sce$CellType)
ct_colours["beta"] <- "yellow"
ct_colours[c("alpha", "delta")] <- "cyan"
ct_colours["Tc"] <- "red"
ct_colours["Th"] <- "blue"

# Plot the cells
plotCells(mask = cur_masks,
          object = cur_sce,
          cell_id = "CellNumber",
          img_id = "ImageName",
          colour_by = "CellType",
          colour = list(CellType = ct_colours),
          image_title = list(text = c("Non-diabetic",
                                      "Recent onset T1D",
                                      "Long duration T1D"),
                             colour = "black"),
          scale_bar = list(length = 100,
                           label = expression("100 " ~ mu * "m"),
                           colour = "black"),
          missing_colour = "white",
          background_colour = "gray")

# Save figure
plotCells(mask = cur_masks,
          object = cur_sce,
          cell_id = "CellNumber",
          img_id = "ImageName",
          colour_by = "CellType",
          colour = list(CellType = ct_colours),
          image_title = list(text = c("Non-diabetic",
                                      "Recent onset T1D",
                                      "Long duration T1D"),
                             colour = "black"),
          scale_bar = list(length = 100,
                           label = expression("100 " ~ mu * "m"),
                           colour = "black"),
          missing_colour = "white",
          background_colour = "gray", 
          save_plot = list(filename = "docs/final_figures/main/Fig_2A.png", scale = 3))
```

We now see that T cells are located closer to the islet on the "Recent onset T1D" image. 
In addition, no beta cells are visible on the "Long duration T1D" image, as expected because they have been destroyed by the immune system.    

We will now confirm that this is reflected on the original images by (i) plotting the corresponding expression values on the mask using the `plotCells` function and (ii) plotting the pixel-level data with the `plotPixels` function. 
CD4 (`CD4` expressed by Th cells) will be plotted in blue, CD8 (`CD8a`, Tc cells) in red and proinsulin (`PIN`, beta cells) in yellow.

First, we will plot the arcsinh-transformed averaged expression on the images.

```{r plotCells-meanExpression}
plotCells(mask = cur_masks,
          object = cur_sce,
          cell_id = "CellNumber",
          img_id = "ImageName",
          exprs_values = "exprs",
          colour_by = c("PIN", "CD4", "CD8a"), 
           colour = list(PIN = c("black", "yellow"),
                         CD4 = c("black", "blue"),
                         CD8a = c("black", "red")),
          image_title = list(text = c("Non-diabetic",
                                      "Recent onset T1D",
                                      "Long duration T1D"),
                             colour = "black"),
          scale_bar = list(length = 100,
                           label = expression("100 " ~ mu * "m"),
                           colour = "black"),
          missing_colour = "white",
          background_colour = "gray")

# Save figure
plotCells(mask = cur_masks,
          object = cur_sce,
          cell_id = "CellNumber",
          img_id = "ImageName",
          exprs_values = "exprs",
          colour_by = c("PIN", "CD4", "CD8a"), 
           colour = list(PIN = c("black", "yellow"),
                         CD4 = c("black", "blue"),
                         CD8a = c("black", "red")),
          image_title = list(text = c("Non-diabetic",
                                      "Recent onset T1D",
                                      "Long duration T1D"),
                             colour = "black"),
          scale_bar = list(length = 100,
                           label = expression("100 " ~ mu * "m"),
                           colour = "black"),
          missing_colour = "white",
          background_colour = "gray",
           save_plot = list(filename = "docs/final_figures/main/Fig_2B.png", scale = 3))
```

Finally, we will display the pixel-level information on the images.

```{r plotPixels}
plotPixels(image = cur_images,
           cell_id = "CellNumber",
           img_id = "ImageName",
           colour_by = c("PIN", "CD4", "CD8a"), 
           colour = list(PIN = c("black", "yellow"),
                         # SYP = c("black", "cyan"),
                         CD4 = c("black", "blue"),
                         CD8a = c("black", "red")),
           bcg = list(PIN = c(0, 4, 1),
                      # SYP = c(0, 3, 1),
                      CD4 = c(0, 3, 1),
                      CD8a = c(0, 3, 1)),
           image_title = list(text = c("Non-diabetic",
                                       "Recent onset T1D",
                                       "Long duration T1D")),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m")))

# Save figure
plotPixels(image = cur_images,
           cell_id = "CellNumber",
           img_id = "ImageName",
           colour_by = c("PIN", "CD4", "CD8a"), 
           colour = list(PIN = c("black", "yellow"),
                         # SYP = c("black", "cyan"),
                         CD4 = c("black", "blue"),
                         CD8a = c("black", "red")),
           bcg = list(PIN = c(0, 4, 1),
                      # SYP = c(0, 3, 1),
                      CD4 = c(0, 3, 1),
                      CD8a = c(0, 3, 1)),
           image_title = list(text = c("Non-diabetic",
                                       "Recent onset T1D",
                                       "Long duration T1D")),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m")),
           save_plot = list(filename = "docs/final_figures/main/Fig_2C.png", scale = 3))
```

This plot confirms that beta cells are surrounded by infiltrating T cells on the "Recent onset T1D" image. 
In addition, proinsulin abundance seems to be reduced on that image as compared to the "Non-diabetic" control image. 
This observation can be further explored by performing additional data analysis.

## Figure 3
