==> analysis/04-Figure_S3.Rmd <==
---
title: "Supplementary Figure 3"
author: "Nils Eling and Nicolas Damond"
date: "`r Sys.Date()`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script reproduces the analysis performed in Supplementary Figure 3.
Here, we will load the libraries and data for this figure:

```{r load-libraries-and-data, message=FALSE}
library(cytomapper)
library(dplyr)

sce <- readRDS("data/PancreasData/pancreas_sce.rds")
masks <- readRDS("data/PancreasData/pancreas_masks.rds")
images <- readRDS("data/PancreasData/pancreas_images.rds")
```

Here, we will highlight a few images that contain different cell-types and outline these using the segmentation masks.
This analysis will visually confirm cell-type phenotyping and segmentation results.

We will first select images with a high count of CD4 and CD8 T cells.

```{r select-Tcells}
# Select the image with the highest T cell density
selected_images <- as_tibble(colData(sce)) %>% 
  # Calculate for each image the area, number of T cells and T cell density
  group_by(ImageName) %>%
  summarise(width = mean(width),
            height = mean(height),
            ImageArea = (width * height) / 10^6,
            TcellCount = sum(CellType == "Tc" | CellType == "Th"),
            TcellDensity = TcellCount / ImageArea) %>%
  arrange(desc(TcellDensity))
```

Now, we will visualize the top image and outline CD4 and CD8 T cells.

```{r highlight-Tcell-outline}
top_images <- selected_images$ImageName[1]

cur_images <- images[match(top_images, mcols(images)$ImageName)]
cur_masks <- masks[match(top_images, mcols(images)$ImageName)]
cur_sce <- sce[,sce$CellType == "Th"]

plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "CD4"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         CD4 = c("black", "red"),
                         CellType = c(Th = "white")),
           bcg = list(H3 = c(0, 1.5, 1),
                      CD4 = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(20, 20)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50))

# Save image
plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "CD4"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         CD4 = c("black", "red"),
                         CellType = c(Th = "white")),
           bcg = list(H3 = c(0, 1.5, 1),
                      CD4 = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(20, 20)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50),
           save_plot = list(filename = "docs/final_figures/supplements/Fig_S3A.png", scale = 3))

cur_sce <- sce[,sce$CellType == "Tc"]

plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "CD8a"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         CD8a = c("black", "red"),
                         CellType = c(Tc = "white")),
           bcg = list(H3 = c(0, 1.5, 1),
                      CD8a = c(0, 6, 1)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(20, 20)))

# Save image
plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "CD8a"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         CD8a = c("black", "red"),
                         CellType = c(Tc = "white")),
           bcg = list(H3 = c(0, 1.5, 1),
                      CD8a = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(20, 20)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50),
           save_plot = list(filename = "docs/final_figures/supplements/Fig_S3B.png", scale = 3))
```

As a second example, we will select images with high alpha and beta cell count and perform a similar analysis as above.
Due to the loss of beta cells, we will only select images of healthy patients.

```{r select-islet-cells}
# Select the three images with the highest alpha and beta cell density
selected_images <- as_tibble(colData(sce)) %>% 
  filter(stage == "Non-diabetic") %>%
  group_by(ImageName) %>%
  summarise(width = mean(width),
            height = mean(height),
            ImageArea = (width * height) / 10^6,
            alphaCellCount = sum(CellType == "alpha"),
            alphaCellDensity = alphaCellCount / ImageArea,
            betaCellCount = sum(CellType == "beta"),
            betaCellDensity = betaCellCount / ImageArea) %>%
  mutate(alphaCellRank = rank(alphaCellDensity),
         betaCellRank = rank(betaCellDensity),
         rankSum = alphaCellRank + betaCellRank) %>%
  arrange(desc(rankSum))
```

We will now outline alpha and beta cells.

```{r outline-alphaAndBeta-cells}
top_images <- selected_images$ImageName[1]

cur_images <- images[match(top_images, mcols(images)$ImageName)]
cur_masks <- masks[match(top_images, mcols(images)$ImageName)]
cur_sce <- sce[,sce$CellType == "alpha"]

plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "GCG"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         GCG = c("black", "red"),
                         CellType = c(alpha = "white")),
           bcg = list(H3 = c(0, 6, 1),
                      GCG = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(40, 40)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50))

# Save image
plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "GCG"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         GCG = c("black", "red"),
                         CellType = c(alpha = "white")),
           bcg = list(H3 = c(0, 6, 1),
                      GCG = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(40, 40)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50),
           save_plot = list(filename = "docs/final_figures/supplements/Fig_S3C.png", scale = 3))

cur_sce <- sce[,sce$CellType == "beta"]

plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "PIN"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         PIN = c("black", "red"),
                         CellType = c(beta = "white")),
           bcg = list(H3 = c(0, 6, 1),
                      PIN = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(40, 40)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50))

# Save image
plotPixels(image = cur_images,
           object = cur_sce,
           mask = cur_masks,
           img_id = "ImageName",
           cell_id = "CellNumber",
           colour_by = c("H3", "PIN"),
           outline_by = "CellType",
           colour = list(H3 = c("black", "blue"),
                         PIN = c("black", "red"),
                         CellType = c(beta = "white")),
           bcg = list(H3 = c(0, 6, 1),
                      PIN = c(0, 6, 1)),
           scale_bar = list(length = 100,
                            label = expression("100 " ~ mu * "m"),
                            margin = c(40, 40)),
           legend = list(colour_by.title.cex = 1.5,
                         margin = 50),
           save_plot = list(filename = "docs/final_figures/supplements/Fig_S3D.png", scale = 3))
```
